separate kintoneModel to structure and model
